Push value validators into Value namespace to shorten class names.
