+++ lib/Validator/Value/MultiPointValidator.php
@@ -21,48 +21,23 @@
  * SOFTWARE.
  */
 
-namespace CrEOF\Geo\Obj\Validator;
+namespace CrEOF\Geo\Obj\Validator\Value;
 
-use CrEOF\Geo\Obj\Exception\ExceptionInterface;
-use CrEOF\Geo\Obj\Exception\RangeException;
-use CrEOF\Geo\Obj\Exception\UnexpectedValueException;
 use CrEOF\Geo\Obj\Object;
 
 /**
- * Class PointValueValidator
+ * Class MultiPointValidator
  *
  * @author  Derek J. Lambert <dlambert@dereklambert.com>
  * @license http://dlambert.mit-license.org MIT
  */
-class PointValueValidator extends AbstractValidator
+class MultiPointValidator extends LineStringValidator
 {
     /**
-     * PointValueValidator constructor
+     * MultiPointValidator constructor
      */
     public function __construct()
     {
-        $this->expectedType = Object::T_POINT;
-    }
-
-    /**
-     * @param array $value
-     *
-     * @throws ExceptionInterface
-     */
-    public function validate(array $value)
-    {
-        parent::validate($value);
-
-        $count = count($value['value']);
-
-        if ($count < 2 || $count > 4) {
-            throw new RangeException('Point value count must be between 2 and 4.');
-        }
-
-        foreach ($value['value'] as $num) {
-            if (! is_int($num) && ! is_float($num)) {
-                throw new UnexpectedValueException('Point value must be array containing "integer" or "float", "' . gettype($num) . '" found');
-            }
-        }
+        $this->expectedType = Object::T_MULTIPOINT;
     }
 }
